backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
pymysql==1.1.2
bcrypt==5.0.0
PyJWT==2.12.0
python-dotenv==1.2.2
pytest==9.0.2
httpx==0.28.1
